Insert data only from orq master
