+++ lib/orquestadores/orquestador_dynamic_supervisor.ex
@@ -9,10 +9,10 @@ defmodule OrquestadorDynamicSupervisor do
     DynamicSupervisor.init(strategy: :one_for_one, max_restarts: 5, max_seconds: 5)
   end
 
-  def start_child(child_name) do
-    spec = {Orquestador, child_name}
+  def start_child(orchestrator_id, type) do
+    spec = {Orquestador, {orchestrator_id, type}}
     DynamicSupervisor.start_child(__MODULE__, spec)
-  end
+   end
 end
 
 # Examples
